# Improve performance of `LineLengthRule`