This depends on https://github.com/jpsim/SourceKitten/pull/152
The duration of linting Carthage 0.11 is reduced from 17sec to 16sec
SwiftLint(32b325b) with SourceKitten(0f54e19):
```
swiftlint lint --config ~/.swiftlint-test.yml  17.48s user 1.12s system 83% cpu 22.182 total
```
SwiftLint(this) with SourceKitten(d7f9266):
```
swiftlint lint --config ~/.swiftlint-test.yml  16.31s user 1.05s system 85% cpu 20.313 total
```
On Instruments, `LineLengthRule` are reduced from:
<img width="781" alt="screenshot 2016-01-30 14 12 44" src="https://cloud.githubusercontent.com/assets/33430/12693843/f836afee-c75b-11e5-985f-e1c85614ca53.png">
to:
<img width="781" alt="screenshot 2016-01-30 14 13 35" src="https://cloud.githubusercontent.com/assets/33430/12693845/fff9824c-c75b-11e5-90c9-5642bf072e8e.png">