@norio-nomura 2016-01-30
Rebased.