@jpsim 2016-01-31
I'll rebase and merge after #452 is in. Thanks @norio-nomura!